feat(pollution_vehicules_eu): added graph for pollution vehicules in eu, refacto name of projet
