+++ SG_AH_pollution_des_transports/get_data.py
@@ -31,22 +31,11 @@ def transform_countries_names(df, col):
                        'RO' : 'Roumanie', 'SE' : 'Suède', 'SI' : 'Slovénie',
                        'SK' : 'Slovaquie', 'TR' : 'Turquie', 'UK' : 'Royaume-Uni'}
 
-    countries_name_en = { 'AT': 'Austria', 'BE' : 'Belgium', 'BG' : 'Bulgaria',
-                       'CH' : 'Switzerland', 'CY': 'Cyprus', 'CZ': 'Czech Republic',
-                       'DE' : 'Germany', 'DK' : 'Denmark', 'EE' : 'Estonia',
-                       'EL' : 'Greece', 'ES' : 'Spain', 'FI' : 'Finland',
-                       'FR' : 'France', 'HR' : 'Croatia', 'HU' : 'Hungary',
-                       'IE' : 'Ireland', 'IS' : 'Iceland', 'IT' : 'Italy',
-                       'LI' : 'Liechtenstein', 'LT' : 'Lithuania', 'LU' : 'Luxembourg',
-                       'LV' : 'Latvia', 'MT' : 'Malta', 'NL' : 'Netherlands',
-                       'NO' : 'Norway', 'PL' : 'Poland', 'PT' : 'Portugal',
-                       'RO' : 'Romania', 'SE' : 'Sweden', 'SI' : 'Slovenia',
-                       'SK' : 'Slovakia', 'TR' : 'Turkey', 'UK' : 'United Kingdom'}
     df[col].replace(countries_name, inplace = True)
     return df
 import os
 def get_transport_pollution_eu():
-    df = pd.read_csv("SG_AH_emission_de_CO2_des_transports/data/emission_de_polluant_des_transports_1990-2020.csv")
+    df = pd.read_csv("SG_AH_pollution_des_transports/data/emission_de_polluant_des_transports_1990-2020.csv")
     
     # Remove the useless columns, keeping only:
     # The data (TIME_PERIOD), The country (geo), Pollution type (airpol), the value of the pollution (OBS_VALUE) 
@@ -62,8 +51,8 @@ def get_transport_pollution_eu():
     df = transform_countries_names(df, 'Pays')
     return (df_all_eu, df)
 
-def get_mean_co2_new_vehicules():
-    df = pd.read_csv("SG_AH_emission_de_CO2_des_transports/data/moyenne_emissions_CO2_vehicules_neuf_2000-2020.csv")
+def get_pollution_per_vehicules_eu():
+    df = pd.read_csv("SG_AH_pollution_des_transports/data/moyenne_emissions_CO2_vehicules_neuf_2000-2020.csv")
 
     # Remove the useless columns
     # Keep : geo, TIME_PERIOD, OBS_VALUE
@@ -72,7 +61,8 @@ def get_mean_co2_new_vehicules():
     df = df.rename(columns= {'geo': 'Pays', 'TIME_PERIOD': 'Année', 'OBS_VALUE' :'Taux de pollution'})
 
     # Seperate data for all EUROPE
-    df_all_eu = df.loc[(df['Pays'] == 'EU27_2020')]
+    df_all_eu = df.copy().loc[(df['Pays'] == 'EU27_2020')]
+    df_all_eu['Pays'].replace({'EU27_2020' : 'Europe Entière', 'EU27_2007' : 'Europe Entière', 'EU28' : 'Europe Entière'}, inplace=True)
     df = df[(df['Pays'] != 'EU27_2020') & (df['Pays'] != 'EU27_2007') & (df['Pays'] != 'EU28')]
     
     # Rename geo columns with right names
@@ -80,7 +70,7 @@ def get_mean_co2_new_vehicules():
     return (df, df_all_eu)
 
 def get_air_pollution_schools():
-    df = pd.read_csv("SG_AH_emission_de_CO2_des_transports/data/ecoles-creches-idf-prepared_2012-2017.csv")
+    df = pd.read_csv("SG_AH_pollution_des_transports/data/ecoles-creches-idf-prepared_2012-2017.csv")
 
     # remove useless columns
     df = df.drop(['ID', 'ville', 'CP', 'type', 'geometry'], axis=1)
@@ -93,7 +83,7 @@ def get_air_pollution_schools():
     return (df_no2, df_pm10, df_pm25)
 
 def get_pollution_per_vehicules_in_france():
-    df = pd.read_csv("SG_AH_emission_de_CO2_des_transports/data/vehicules_polluant_france_2015.csv",sep=";",encoding="latin1")
+    df = pd.read_csv("SG_AH_pollution_des_transports/data/vehicules_polluant_france_2015.csv",sep=";",encoding="latin1")
 
     # Remove the useless columns
     # Keep : geo, TIME_PERIOD, OBS_VALUE
